Long live QArrayAllocatedData
QArrayAllocatedData is the actual base class of the allocated data for
QString, QByteArray and QVector instead of QArrayData. It builds upon
QArrayData and moves the allocated size there.

QArrayData itself remains POD and is used for all static, read-only
purposes, like Q_ARRAY_LITERAL and QStringLiteral. It has, however,
one member fewer, which means its size decreases from 20 bytes
(32-bit) / 32 bytes (64-bit) to 16 / 24 bytes. This brings the size of
statics back to what it was prior to the addition of the flags
field. Dynamic memory allocations continue consuming 20 / 32 bytes.

As a side-effect, every call to capacity() checks the type flag. Such
calls should be minimised and should be confined only to places where
we're considering reallocation anyway. In particular it should not
happen in read-only contexts.

Change-Id: Ib2d7e78e3494d827adb6c3f12ba48d4eb93f1ab5