Remove the reference counter from static QArrayData objects
The reference counter for static objects cannot be changed anyway, so
we don't need to store it. This allows us to reduce the overhead on
QByteArrayLiteral and QStringLiteral to the sweet spot of 16 bytes.

However, as an unintended consequence, all static objects now report
that they're shared (isShared() == true), including the unsharable
empty optimisation. Technically, the information is correct: it *is*
being shared.

Change-Id: I187ee12053b941e1a6b021e783ebeb3caa4439d1