Introduce an adaptative mutex locking algorithm
Like in Qt 4 days and like pthread_mutex type PTHREAD_MUTEX_ADAPTIVE_NP,
spin for a while like a spinlock before we give up and try to
sleep. This is enabled only for when a multi-processor system is
detected -- there's no point in spinning on a single-processor, as we
may not lose our CPU timeslice soon.

When running with one CPU, there's no noticeable change in
performance. QMutex performance is also equivalent to pthread_mutex_t
performance.

When running with 4 cores, on my Intel Sandybridge system, I get

                 tick count * 1000        CPU time (µs)
                1 mutex   2 mutexes    1 mutex  2 mutexes
pthread           1462      2762         3735      6334
QMutex before     1210      1994         3076      4771
QMutex after      1115      1319         2878      3422
 after/pthread     -23.7%    -52.2%       -23.0%    -46.0%
 after/before       -7.8%    -33.8%        -6.4%    -28.3%

The interpretation is that the operation takes less time in all and less
CPU power, but the process is active (running) for slightly longer.

Change-Id: I88b436c5d784bf742eaa46964d0195a2328fd20e