Add a function to run the PAUSE instruction on x86
According to Intel's optimisation manual, a processor looping very fast
trying to acquire a spinlock or some other primitive will flood the
memory bus and cache with requests, making it impossible for other
threads (on the same core) or other cores from releasing the resource in
question.

Therefore, inserting a PAUSE improves performance. This function does
that.

Also note that PAUSE causes a transaction abort if a transaction is
currently running. Having PAUSE close to the beginning of a contention
will make the processor abort sooner and retry without transactions.

Change-Id: Ie40dcd0b3bce8d8f378475f44b49a2873ccda732