Rewrite the QMutex lock and unlock code on without Linux futexes
This solution uses a spinlock on the d pointer's lowest bit to protect
the updating of the reference counter. This solves a potential race
condition in the previous code as well as remove the need for the
"possiblyUnlocked" trick.

The refCount member now always contains the exact number of threads
contending for this mutex, which means that when it drops to 1 (and it
can only be dropped while spinlocked), the d pointer can be released.

Change-Id: I310e20ebb2b45b6a4f0ac4f723cac4f5cea97dae