Add QArrayData::sharedNullData()
Just to simplify a few operations, like detecting when a QChar* or char*
coming from a QString or QByteArray, respectively, were null data. Start
moving QArrayData's size and data pointer to the main class

While you're not supposed to dereference the pointer returned by
QVector::data() unless you know that the array is non-empty, that is
permitted for QString and QByteArray. That is, QString().constData()
must return a valid pointer to a null QChar.

Change-Id: I80b4b62f203dc841e5c99c20c51d92ca576e4bfe