Add/update benchmarks for several QString algorithms
=== Latin1 ===

The SSE2 and Neon implementations have been there since Qt 4.7; the
SSE4.1 and 4.2 updates were added in Qt 4.8. During Qt 4.7 development,
we benchmarked the implementations and found that they improved
performance (the comments say so). However, the SSE4.1 and 4.2
implementations were not benchmarked and the benchmark shows the SSE4.2
code is not an improvement.

Results of the benchmarks (improvement over the plain version):

        Creator startup Creator startup + load session
SSE2         20.4%                      76.1%
SSE4.1       24.5%                      77.4%
SSE4.2       12.7%                      57.6%

=== UTF-8 codec ===

The code being added is copied from qutfcodec.cpp in QtCore.  It's a
copy for two reasons:
 1) it's static, so it won't be affected by changes in QtCore. We need
    it so we can run the comparison against the codec that was there
    in Qt 4.7 through 5.2)

 2) it's slightly modified so it doesn't return a QString or
    QByteArray.  We do that to avoid malloc() in the benchmark loop,
    which is rather non deterministic.

The results of this benchmark were run with data gathered from a run
of QtCreator (start application, load pre-existing session containing
many Qt 5 libraries and unit tests). All calls to toUtf8 and fromUtf8
were logged. The data is not included in the commit because it's too
large, only a small sample version is included.

The improvement for the non-SIMD version, measured in both CPU ticks and CPU
cycles (best of 10 runs) on my SandyBridge machine running at 2.66 GHz (turbo
@ 3.3 GHz), is:

                  Compared to old codec           Compared to ICU
Data              fromUtf8       toUtf8         fromUtf8       toUtf8
ASCII only         1.55x         2.29x           1.67x         1.87x
UTF-8 2-char       1.97x         1.74x           1.41x         1.53x
UTF-8 3-char       1.89x         1.51x           1.38x         1.40x
UTF-8 4-char       1.50x         1.58x           1.31x         1.46x
Data from Creator  1.38x         2.35x           2.79x         1.72x

The data extracted from Creator was mostly ASCII, with a median and
mode length of 12 characters, but with a few very, very long strings
(350k characters).

=== UTF-8 codec with SIMD ===

Decoding from UTF-8 is easy: if the high bit is set, we fall back to
the byte-by-byte decoding. Encoding to UTF-8 requires a little bit
more work: to detect anything between 0x0080 and 0xffff, we have
several options but none as easy as above. Multiple alternatives are
in the benchmark code.

In both loops, we do two things once we run into a non-ASCII
character: first, we continue the loop for the remainder of ASCII
characters in the buffer (which we can tell by checking the bits set
in the mask), then we find the last non-ASCII character in that
16-character group, so we don't reenter the SSE code too soon.

For the UTF-8 encoding, I have chosen the alternative that results in
the best performance. It's closely tied to the alternative running the
PMIN instruction, but that requires SSE 4.1. It's not worth the
complexity. And quite counter-intuitively, the dedicated string
instruction from SSE 4.2 performs most poorly of all solutions. This
begs re-visiting the performance of the toLatin1 encoder.

The best of 10 benchmark runs of this code were measured on my
SandyBridge CPU @ 2.66 GHz (turbo @ 3.3 GHz), both as CPU cycles and
as CPU ticks:

  Compared to:        ICU             Qt 4.7      non-SSE Qt 5.3
  Data set     fromUtf8  toUtf8  fromUtf8 toUtf8  fromUtf8 toUtf8
 ASCII only      7.50x    6.22x    6.94x   7.60x    4.45x   4.90x
2-char UTF-8     1.17x    1.33x    1.64x   1.56x    1.01x   1.02x
3-char UTF-8     1.08x    1.18x    1.48x   1.33x    0.97x   0.92x
4-char UTF-8     1.05x    1.19x    1.20x   1.21x    0.97x   0.97x
Creator data     3.62x    2.16x    2.60x   1.25x    1.78x   1.23x

As shown by the numbers, the SSE-based code is slightly worse than the
non-SSE code for dense non-ASCII strings. However, as evident in the
Qt Creator data, most strings manipulated by applications are either
pure ASCII or mostly so, so there's a net gain.

Done-with: H. Peter Anvin <hpa@linux.intel.com>

=== ucstrncmp ===

This also removes the QString benchmark for qMemEquals since ucstrncmp
and qMemEquals are being merged into one single function. That merge
alone shows an improvement that is worthwhile, even without the
further SIMD improvements:

                GCC 4.7         GCC 4.9         Clang 3.4       ICC 14
"intwise"        22.7%           24.0%            22.8%          31.6%
"sse2"           54.6%           47.4%            41.5%          52.9%

Next, we compared the basic "sse2" function with more complex examples
trying to squeeze out more performance. In all cases, the best
improvement was only 3.8% over the basic version and the results
aren't consistent between the compilers.

The benchmark was run on a Sandybridge Core-i7 2620M @ 2.7 GHz, with
turbo mode disabled. The results are the best of 10 runs (not
average).

=== findChar ===

This updates only the case-sensitive searching. The case-insensitive
part requires quite a few Unicode transformations.

The benchmarks tried are the plain word-by-word comparison used in Qt
5.2 and in builds without SSE2; the SSE2 benchmark; a benchmark using
the SSE4.2 "strchr" instruction. I've run the benchmark both for CPU
cyles used as well as bytes/sec scanning strings.

Improvement over the Qt 5.2 code:

            GCC 4.7         GCC 4.9        Clang 3.4        ICC 14
          cycles  MB/s    cycles  MB/s    cycles  MB/s    cycles  MB/s
SSE2       2.1x   2.9x     2.2x   2.9x     2.1x   3.1x     2.2x   3.1x
SSE4.2     1.5x   1.7x     1.5x   1.7x     1.5x   1.7x     1.6x   1.8x

Once again, the SSE4.2 instruction wasn't as effective as I'd hoped
(not to mention that Clang seems to have some bugs emitting it).

Change-Id: I2308761ab951120122005bb00bbf7ec9802a8b53