Don't crash when detaching a capacity-reserved arrays
When an array needs to be detached, it is either because we're out of
capacity or because we're about to modify a container that is
currently shared. If it's the latter, we should not pass the
Grows{Forward,Backwards} flags.

Without this patch, it's easy to make a container crash in about 30
iterations:
  QVector<int> copy;
  for (int i = 0; i < 31; ++i) {
      copy = vector;
      vector.append(i);
  }

By the end of this loop, vector.size() == 32 but the vector has
allocated 2 GB of data.