Handle insert/prepend/append of items in the current vector
If the items being inserted/copied come from the current vector, we
need to either copy the item or we need to copy onto a new
vector. Since append() and prepend() are implemented in terms of
ranges, we end up with a small pessimization of the code: we copy the
rest of the vector, instead of the one item being inserted.

Change-Id: I91ed55de8813b4b8b2c8b70dba7cb0ce84fa381a