WIP Starting using AVX512F
AVX512F continues the trend by adding the PMOVZXBD instruction that
extends from one 128-bit register to a 512-bit register. Coupled with
PMOVZXWD to extend from 256 to 512 bits, we can then do the comparison
and use the new mask registers.

For the other operations, AVX512F is going to be interesting with
32-character loops only.

Change-Id: I60f4fbb7734c017e1411ae8498020f36dbc1ee28