Add support for char16_t and char32_t to QVariant
This commit adds the necessary numeric comparison hooks for char16_t and
char32_t. Mostly, those are treated simply the same as quint16 and
quint32. The only difference is when converting from those to QString,
in which case we treat them as characters, not numbers.

This compiles in both C++98 and C++11 modes. The C++98 mode support is
provided by having structs with the name of the C++11 types (which may
case some compilers to print warnings, but you should instead compile in
C++11 mode).

[ChangeLog][QtCore][QMetaType and QVariant] Added support for char16_t and
char32_t to the Qt metatype system.

Change-Id: I7409ee2757e2120104b41351e9312bcd406f14fa