Fix the reference counting of QDBusArgumentPrivate
It was a mess up until now. It was refcounted, but the decision on
whether to detach or not was based on a combination of the reference
counter and the presence of a DBusMessage*. That solution for handling
the lifetime made it a little harder to make QDBusMarshaller own a
DBusMessage of its own, which the next commit will require.

Instead, use -1 for a "static" reference (like QStringLiteral does).
Those references are the QDBusMarshaller and QDBusDemarshaller that are
created on the stack by QtDBus itself during marshalling and
demarshalling. The objects with ref == 1 are those that we "dup" during
demarshalling or created by the user by the QDBusArgument() constructor.

Change-Id: Ic5d393bfd36e48a193fcffff13b7375a4328a811