Add support for O_NOSIGPIPE and SOCK_NOSIGPIPE
When those flags are set, the kernel will not send SIGPIPE to writes to
empty pipes or disconnected sockets. So we don't need to set SIGPIPE to
SIG_IGN and possibly interfere with legitimate use of it.

Currently, after the call to qt_ignore_sigpipe, we store a global in
QtCore that indicates what we've ignored it. We use the same variable
now to store whether for pipe2, socket and accept4 support O_NOSIGPIPE /
SOCK_NOSIGPIPE. This also assumes that if one of them supports the flag,
then all do.

This flag is supported on NetBSD and a patch has been sent for Linux.

Change-Id: Iee8cbc07c4434ce9b560ffff13cb84d8888ac803