Make QCoreApplication abort execution if it detects copy relocations
A copy relocation is a type of ELF relocation where the entire variable
is moved from one place to another. Before executing any code, the
linker will find where the actual variable is located, then copy its
contents to the new place (hence the name), and point everyone there.
Those relocations are a problem if Qt was compiled with the configure
option -reduce-relocations.

The comparison should never fail on non-ELF systems or on systems where
-reduce-relocations was disabled, but leave it there anyway.

Task-number: QTBUG-45755
Change-Id: I66a35ce5f88941f29aa6ffff13dd9380cb67e99f