Fix support for QCOMPARE'ing function pointers
They are regular pointers, except that they don't automatically convert
to void*. So we can't call a function that would require that conversion
(compare_ptr_helper).

So I removed that function and instead used the fact that they don't
cast automatically to void* to check if the conversion can happen.

Finally, this required rearranging some functions a little to make sure
that the correct definitions are visible at time of the template
declarations.

Change-Id: I4078d5b5bbe246048674ffff14156ef28e37b44a