QtTest: Add support for printing pointer-to-member functions with GCC
With the IA-64 C++ ABI, the format of a pointer-to-member function is
actually well-defined and we can easily extract the actual function
pointed to from it, if it's not virtual. For virtual functions, we'd
need a pointer to a concrete object to extract the vtable from and we
don't have that.

GCC, Clang and several other compilers when in compatibility mode with
those two will always use the IA-64 C++ ABI. There are some other
compilers that also use it and don't define __GNUC__ (such as HP's aCC
for Itanium), but Qt isn't currently tested on them, so support isn't
added in this commit.

Change-Id: I461e9fc7199748faa187ffff1415d4ebd6264b53