Make moc use the actual preprocessor
Problems found so far:
* if QMAKE_MOC_OPTIONS is set, we can't use the preprocessor
  -> this explains why tst_Moc::specifyMetaTagsFromCmdline is failing
  I don't see a solution, aside from the isEmpty() in moc.prf

* Q_GADGET processing is broken because actualSource() skips all the
  other files
  -> tst_Moc::relatedMetaObjectsInGadget is failing

* line numbers in errors and warnings are wrong

For the last one, there are two solutions: either insert blanks to keep
the right count or teach the tokeniser to understand the line number
information.

For the Q_GADGET, the solution will be to make the preprocessor *not*
drop the content of the other files, which means we will need to teach
Moc itself about the # lines.

Change-Id: Id75834dab9ed466e94c7ffff14455a562c4ca34d