Implement futexes for Windows 8
Windows 8 added this pair of functions that can be used to implement the
same functionality as we have on Linux. For ease of understanding, I'm
calling them "futex" on Windows too.

Also note that the minimum version of WinRT applications is Windows 8,
so we can link directly to the functions in question.

Change-Id: Ifea6e497f11a461db432ffff1448c6806ecfc36c