Implement coarse QDeadlineTimer for Unix (Linux)
Linux has CLOCK_MONOTONIC_COARSE, with a precision of 4 ms, which is
much faster to compute since it only accesses the "current time" data
that is shared among all processes, with no access to timing hardware.

I had to adapt certain other functions in order to make this
happen. Where qt_gettime() was called, the choice for Qt::CoarseTimer is
whether the timeout being handled is 2 s or more (same logic as
QTimer::singleShot). In the case of the event loop, it depends on the
first non-active timer, since the list is already sorted anyway.

Change-Id: I25d85d86649448d5b2b3fffd1450bffa711f9604