moc: add a feature to parse all of the headers in one go
This launches moc only once and, though it will parse all the headers
multiple times (we don't have a precompiled header support), it's still
faster than not.

On my machine (i7-6560U), the time to run moc all QtCore headers is:
    9,34s user 2,22s system 368% cpu 3,137 total  [just moc]
    25,88s user 4,17s system 369% cpu 8,139 total [moc + compile]

with this patch, that time reduces to:
    5,81s user 0,88s system 99% cpu 6,744 total   [just moc]
    7,07s user 0,96s system 101% cpu 7,923 total  [moc + compile]

In both cases, there's a sharp decrease in CPU time, even if moc's
single-threadedness means the wall-clock time is higher for just moc'ing
the QtCore headers. The important fact is that more CPU time is free now
for compiling the rest of the source code.

[ChangeLog][moc] Moc is now able to parse multiple headers at the same
time and produce a single output, which reduces build time. This
functionality is enabled if the --combine flag is passed on the
command-line. For qmake-based projects, this feature can be enabled by
using CONFIG += moc_combine. Note that this feature is incompatible with
Q_PRIVATE_SLOT(), with precompiled headers that #include the headers of
the module being compiled, and will require removal of any #include for
moc outputs.
\
For other buildsystems, simply pass the new flag to moc and list all
moc'able headers on the command-line or response file.  To improve build
time even further when using the --combine option, it is recommended to
pass -include followed by a header name to execute a "precompilation"
pass.

Change-Id: I87e17314d8b24ae983b1fffd1452430b4f682b3f