QtDBus: remove all but one Q_PRIVATE_SLOT
The one that can't be removed depends on QDBusConection's ability to
connect to a lambda, which it doesn't have yet.

The QDBusServer part required the creation of a QObjectPrivate-derived
class, which didn't exist before (QDBusServer's d pointer was a
QDBusConnectionPrivate, which is a QObject).

The rest are trivial ports. It also made QDBusPendingCallWatcherPrivate
redundant, so it was removed.

Change-Id: I87e17314d8b24ae983b1fffd14531567f1f8b08b