Fix comparison of PMFs in QObjectPrivate::connect() and disconnect()
It is unspecified by the C++ standard if we can compare PMFs of
different classes. Under the IA-64 C++ ABI, the comparison works as
expected for non-virtual members, but it doesn't work for virtual
members. Under the MSVC ABI, it must be similar or possibly worse, due
to the different sizes of PMFs.

Fix this by ensuring we always do a comparison of a QObject member
first.

Task-number: QTBUG-53684
Change-Id: I87e17314d8b24ae983b1fffd145363d54011e851