QObject::connect: Force a conversion to a pointer to QObject member
According to the C++ standard, in 5.10 [expr.eq] Equality Operators,
bullet 3.4, we can only compare two pointer to members if they are
either for the same class or if one of the classes is a base class of
the other. Since all PMF connections are possible only with members of
QObject-derived classes, we can use QObject as that common base.

There's a slight change in code generation, but there's no actual impact
in in performance. Casting a PMF to a base class in the primary chain
leaves it unchanged.

Change-Id: I87e17314d8b24ae983b1fffd145365bef99d7f6e