Add the beginnings of a new-style QDBusConnection::connect
A long time after QObject got it...

This commit adds the machinery necessary for QDBusConnectionPrivate to
handle slot objects. This commit changes the internal SignalHook
structure to carry an extra QtPrivate::SlotObjectBase pointer (whose
refcount we need to manage) and changes all functions dealing with meta
method IDs to pass along that pointer. Finally, deliverCall() gets the
functionality to invoke the stored function.

Change-Id: I87e17314d8b24ae983b1fffd145387f547ec82e4