QtConcurrent: fix calculation of elapsed
The use of double like in difftime(3) is meant to allow differences that
might be larger than 2^63 ticks between uses. First of all, that's
extremely unlikely (it makes sense for difftime, since time_t might be
32-bit).

Second, if you cast the result of the subtraction, you gain nothing. If
it had under- or overflowed, you'll still get bogus values in double.
Instead, cast before the subtraction.

Change-Id: I115db302eb114bed8cd1fffd14557b4397a7dd9b