Make sure we delete QEvents with the event list mutex unlocked
The normal event delivery code already did that (in sendPostedEvents),
but not the clean up of the main thread in cleanupThreadData.

This prevents a deadlock when an event's destructor (usually, a custom
event) causes an event to be posted as a result of its destruction. The
events so posted won't be leaked, as QThreadData's own destructor will
delete them.

Task-number: QTBUG-54282
Change-Id: I2ba169607ed3448d828ffffd145b2713e22d0efe