Filesystem: Use "birth time" to avoid confusion with Unix ctime
The Unix stat fields "st_ctime" and "st_ctim" mean "change time", the
last time that the file/inode status fields were changed. It does not
mean "creation time". So this commit splits all of the internal API to
"birth" and "metadata change" instead of "creation" to avoid the
conflict.

The front-end QFileInfo::created() will be deprecated and replaced in
the next commit.

Change-Id: I149e0540c00745fe8119fffd1463fe78b619649e