WIP Add QDBusAbstractInterface::setProperty and property
These are non-virtual overrides of the QObject functions of the same
name. Prior to Qt 4.6, the code generated by qdbusxml2cpp called
internalPropGet and Set, but in Qt 4.6 it was simplified to call
property() and setProperty(), as we'd rely on the virtual qt_metacall to
do the right thing and dispatch the property getting and setting through
D-Bus.

This commit, by itself, does not do much. But it helps when QObject and
moc start using the qt_static_metacall for getting and setting
properties, as opposed to the current way that uses the virtual
qt_metacall. This change is necessary to break an infinite loop that
would otherwise form: the current solution depended on a base class
handling the properties of a derived class.

The idea is to intercept the calls from the generated qdbusxml2cpp code
for generated properties before they got to the property system.

WIP reason: this regresses two unit tests in tst_QDBusAbstractInterface
and it seems to me the new behaviour should be correct...

Change-Id: I9093948278414644a416fffd1473c5be49166d80