Cause compilation errors if you forget to declare QDataStream operators
The << and >> operators taking a base class of a given type could be
called instead of the real type and we could end up streaming base class
only and only partial information would be saved. This patch adds a
template operator>> and operator<< that return a meaningful struct
indicating that you forgot to add the correct operator.

We were guilty of relying on the automatic casting: for enums (relying
on their cast to int) and QBitMap/QStringList being cast to QPixmap/
QList<QString>. This patch fixes those two issues.

Change-Id: If0ad4d988da143b3b1b2fffd1480d7ab6e758fb3