Replace Qt's hashing function with SipHash
This commit replaces both the standard DJB33XA (DJB's 33 Times Add) and
the faster CRC32 hashing with SipHash. The most important difference
between those algorithms is that both DJB33XA and CRC32 have this
unwelcome property: for two byte sequences x and y, if you know that x
and y have the same hashing for a given seed, then they have the same
hashing for all seeds.

SipHash has no such issue. If the seed changes, the strings that used to
compute to the same hash are no longer likely to do so.

This algorithm is between 1.5x to 4.5x slower than the old DJB33XA
algorithm, and roughly 8x slower than the CRC32 algorithm.

Change-Id: I09100678ff4443e6be06fffd14819c8878d223e2